fixed issues, project owner validation works
